hatchet_worker/hatchet-worker-requirements.txt
python-dotenv
hatchet-sdk==0.26.3
openai
weaviate-client